+++ src/calyxos/utils/debug.py
@@ -2,8 +2,8 @@
 
 from typing import Any
 
-from talos.core.decorator import get_graph
-from talos.graph.node import NodeType
+from calyxos.core.decorator import get_graph
+from calyxos.graph.node import NodeType
 
 
 class GraphDebugger: